You need to hook into both luastates at once.
The VR lua layer is a different lua layer than the flatscreen layer. I noticed your code will check which layer it starts with/finds first and initalizes based on that.

The game defaults to the non-VR mode then turns on the VR mode, so you'll need to hook into both states at the start.

I'd file this as an issue but I can't because no issues enabled on the repo.